*.cpp --> *.c and Debug AI mode
